exemple-projet/pyproject.toml
[tool.poetry]
name = "final"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
rich = "^11.2.0"
pyserde = "^0.7.0"
typer = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
